divide files in assembler
